using-uci-dataset 2/using-uci-dataset/frontend/requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas>=2.0.3
plotly>=5.18.0
extra-streamlit-components==0.1.70
streamlit-option-menu==0.3.12
python-dotenv==1.0.1
fpdf2>=2.7.8
